feat: create splash screen page
